backend/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytils==0.4.1
pytz==2023.4
sqlparse==0.4.4
typing_extensions==4.9.0
